backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.8
